Revert "[#2] refactor: TableHeader Filter,Sort 컴포넌트 분리"
This reverts commit 629a7da5af57934be573e4443fb719b99b1a3f22.
