+++ src/component/common/Table.tsx
@@ -1,6 +1,4 @@
 import { COMMON_COLOR } from './../../constants/colors';
-import TableHeaderFilter from './TableHeaderFilter';
-import TableHeaderSort from './TableHeaderSort';
 import {
   getCoreRowModel,
   useReactTable,
@@ -9,6 +7,8 @@ import {
   getSortedRowModel,
   getFilteredRowModel,
 } from '@tanstack/react-table';
+import { BiSearchAlt } from 'react-icons/bi';
+import { FaSort, FaSortDown, FaSortUp } from 'react-icons/fa';
 import styled, { css } from 'styled-components';
 
 interface TableProps<T extends object> {
@@ -25,6 +25,21 @@ const Table = <T extends object>({ data, columns }: TableProps<T>) => {
     getFilteredRowModel: getFilteredRowModel(),
   });
 
+  const handleFilterStatus = (header: any) => {
+    switch (header.column.getFilterValue()) {
+      case true:
+        return header.column.setFilterValue(false);
+      case false:
+        return header.column.setFilterValue(undefined);
+      case undefined:
+        return header.column.setFilterValue(true);
+    }
+  };
+  const handleSearchName =
+    (header: any) => (e: React.ChangeEvent<HTMLInputElement>) => {
+      header.column.setFilterValue(e.target.value);
+    };
+
   return (
     <TableWrapper>
       <thead>
@@ -32,14 +47,46 @@ const Table = <T extends object>({ data, columns }: TableProps<T>) => {
           <TableRow key={headerGroup.id}>
             {headerGroup.headers.map(header => (
               <Th key={header.id}>
-                {header.isPlaceholder
-                  ? null
-                  : flexRender(
-                      header.column.columnDef.header,
-                      header.getContext()
-                    )}
-                <TableHeaderSort header={header} />
-                <TableHeaderFilter header={header} />
+                <div
+                  {...{
+                    isCanSort: header.column.getCanSort(),
+                    isCanFilter: header.column.getCanFilter(),
+                    onClick: header.column.getToggleSortingHandler(),
+                  }}>
+                  {header.isPlaceholder
+                    ? null
+                    : flexRender(
+                        header.column.columnDef.header,
+                        header.getContext()
+                      )}
+                  {{
+                    asc: <FaSortUp />,
+                    desc: <FaSortDown />,
+                  }[header.column.getIsSorted() as string] ?? null}
+                  {header.column.getCanSort() &&
+                  !header.column.getIsSorted() ? (
+                    <FaSort />
+                  ) : null}
+                </div>
+                {header.column.getCanFilter() ? (
+                  header.id === 'status' ? (
+                    <FilterBtn
+                      onClick={() => handleFilterStatus(header)}
+                      tagValue={
+                        header.column.getFilterValue() as undefined | boolean
+                      }
+                    />
+                  ) : header.id === 'customer_name' ? (
+                    <FilterInput>
+                      <InputBar
+                        type='text'
+                        placeholder='고객 이름 검색'
+                        onChange={handleSearchName(header)}
+                      />
+                      <BiSearchAlt size='17' />
+                    </FilterInput>
+                  ) : null
+                ) : null}
               </Th>
             ))}
           </TableRow>
@@ -115,5 +162,36 @@ const Td = styled.td`
 
   text-align: center;
 `;
-
+const FilterBtn = styled.button<{ tagValue: undefined | boolean }>`
+  outline: none;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  ::before {
+    content: '';
+    display: block;
+    width: 8px;
+    height: 8px;
+    border-radius: 50%;
+    ${({ tagValue }) => {
+      if (tagValue === undefined) return '    border: 1px solid gray;';
+      return tagValue
+        ? 'background-color: #42a6ce'
+        : 'background-color: #e2687c';
+    }};
+  }
+`;
+const FilterInput = styled.div`
+  width: 62%;
+  display: flex;
+  margin: 0 auto;
+  border: 1px solid gray;
+  border-radius: 2rem;
+  padding: 0.3rem 0.7rem;
+`;
+const InputBar = styled.input`
+  width: 100%;
+  outline: none;
+  border: none;
+`;
 export default Table;
